+++ instagram/tests/factories.py
@@ -5,10 +5,10 @@ from factory.alchemy import SQLAlchemyModelFactory
 from factory.fuzzy import FuzzyText
 from faker import Faker
 from instagram import main  # noqa
-from instagram.user.models import InstagramUser, hash_password
-from instagram.follows.models import Follow
+from instagram.follow.models import Follow
 from instagram.like.models import Like
 from instagram.post.models import Comment, Post
+from instagram.user.models import InstagramUser, hash_password
 
 from .database import Session
